+++ src/main/resources/application.properties
@@ -1,5 +1,4 @@
-conductor.server.url=http://localhost:8080/api
+conductor.server.url=https://play.orkes.io/api
 
-conductor.security.client.key-id=96196359-6652-4a3b-9a08-8f2e3b779f0b
-conductor.security.client.secret=gC98Y9KPYhLjZjN34Qe5ZbIFfanCtCH0IL8BIffWBMmjvaTy
-server.port=7000
+conductor.security.client.key-id=_CHANGE_ME_
+conductor.security.client.secret=_CHANGE_ME_
